refactor: split ci identity providers
