+++ libs/zephyr-agent/src/lib/node-persist/ci-token-identity/types.ts
@@ -1,0 +1,17 @@
+export type CiProvider = 'gitlab' | 'github';
+
+export interface CiTokenIdentity {
+  provider: CiProvider;
+  email?: string;
+  emails?: string[];
+  issuer?: string;
+  providerSubject?: string;
+  username?: string;
+  source: 'jwt' | 'api' | 'env' | 'event' | 'noreply';
+}
+
+export interface CiIdentityProvider {
+  provider: CiProvider;
+  detect(env: NodeJS.ProcessEnv): boolean;
+  infer(env: NodeJS.ProcessEnv): Promise<CiTokenIdentity | undefined>;
+}
