+++ libs/zephyr-agent/src/lib/node-persist/ci-token.ts
@@ -1,7 +1,7 @@
 import { StorageKeys } from './storage-keys';
 
 export function getCiToken(): string | undefined {
-  return process.env[StorageKeys.ci_token]?.trim();
+  return process.env[StorageKeys.ze_ci_token]?.trim();
 }
 
 export function hasCiToken() {
